requirement.txt
Flask
msal
azure-storage-blob
python-dotenv